refactor: enforce module boundaries for app, db, and test
Made `app`, `db`, and `testing` internal modules `pub(crate)`.
Exposed required public types like `AppBuilder`, `create_pool`, `TestApp`,
and `TestClient` via explicit `pub use` statements to form a clean facade.
